Create separate methods for handling separate tasks; refs #18513

+++ lib/sitediff/report.rb
@@ -50,34 +50,71 @@ class SiteDiff
       report_before ||= @config.before_url
       report_after ||= @config.after_url
 
-      # Prepare diff directory and wipe out existing diffs.
-      dir = Pathname.new(dir)
-      dir.mkpath unless dir.directory?
-      diff_dir = dir + DIFFS_DIR
-      diff_dir.rmtree if diff_dir.exist?
+      dir = SiteDiff.ensure_dir dir
 
-      # Write diffs to the diff directory.
-      @results.each { |r| r.dump(dir) if r.status == Result::STATUS_FAILURE }
-      SiteDiff.log "All diff files dumped inside #{diff_dir.expand_path}."
+      write_diffs dir
+      write_failures dir
 
-      # Store failing paths to failures file.
-      failures = dir + FAILURES_FILE
-      SiteDiff.log "All failures written to #{failures.expand_path}."
-      failures.open('w') do |f|
-        @results.each { |r| f.puts r.path unless r.success? }
-      end
-
-      # Create report file for each result.
+      # Prepare report.
       report = Diff.generate_html(
         @results,
         report_before,
         report_after,
         @cache
       )
-      dir.+(REPORT_FILE).open('w') { |f| f.write(report) }
 
-      # Prepare settings file for serving reports.
-      # TODO: Find a way to avoid having to create this file.
+      # Write report.
+      report_file = dir + REPORT_FILE
+      report_file.unlink if report_file.file?
+      report_file.open('w') { |f| f.write(report) }
+
+      write_settings dir, report_before, report_after
+    end
+
+    ##
+    # Creates diff files in a directory named "diffs".
+    #
+    # If "dir" is /foo/bar, then diffs will be placed in /foo/bar/diffs.
+    #
+    # @param [Pathname] dir
+    #   The directory in which a "diffs" directory is to be generated.
+    def write_diffs(dir)
+      raise Exception 'dir must be a Pathname' unless dir.is_a? Pathname
+
+      # Delete existing "diffs" dir, if exists.
+      diff_dir = dir + DIFFS_DIR
+      diff_dir.rmtree if diff_dir.exist?
+
+      # Write diffs to the diff directory.
+      @results.each { |r| r.dump(dir) if r.status == Result::STATUS_FAILURE }
+      SiteDiff.log "All diff files written to #{diff_dir.expand_path}"
+    end
+
+    ##
+    # Writes paths with diffs into a file.
+    #
+    # @param [Pathname] dir
+    #   The directory in which the report is to be generated.
+    def write_failures(dir)
+      raise Exception 'dir must be a Pathname' unless dir.is_a? Pathname
+
+      failures = dir + FAILURES_FILE
+      SiteDiff.log "All failures written to #{failures.expand_path}"
+      failures.open('w') do |f|
+        @results.each { |r| f.puts r.path unless r.success? }
+      end
+    end
+
+    ##
+    # Creates report settings.yaml file.
+    #
+    # TODO: Find a way to avoid having to create this file.
+    #
+    # @param [Pathname] dir
+    #   The directory in which the report is to be generated.
+    def write_settings(dir, report_before = nil, report_after = nil)
+      raise Exception 'dir must be a Pathname' unless dir.is_a? Pathname
+
       settings = {
         'before' => report_before,
         'after' => report_after,
@@ -88,7 +125,7 @@ class SiteDiff
 
     ##
     # Returns CSS for HTML report.
-    def css
+    def self.css
       output = ''
       output += File.read(File.join(SiteDiff::FILES_DIR, 'normalize.css'))
       output += File.read(File.join(SiteDiff::FILES_DIR, 'sitediff.css'))
@@ -97,7 +134,7 @@ class SiteDiff
 
     ##
     # Returns JS for HTML report.
-    def js
+    def self.js
       output = ''
       output += File.read(File.join(SiteDiff::FILES_DIR, 'jquery.min.js'))
       output += File.read(File.join(SiteDiff::FILES_DIR, 'sitediff.js'))
